edge cases edit1

+++ CommonCore/libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: hdaoudi <hdaoudi@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 13:48:28 by hdaoudi           #+#    #+#             */
-/*   Updated: 2024/11/07 22:29:37 by hdaoudi          ###   ########.fr       */
+/*   Updated: 2024/11/08 14:16:04 by hdaoudi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	destlen;
-	unsigned int	srclen;
-	
+	size_t	i;
+	size_t	j;
+	size_t	destlen;
+	size_t	srclen;
+
 	i = 0;
 	j = 0;
 	while (dst[j] != '\0')
